refactor(get-platforms): move to get-platforms-sync, to be deprecated
