+++ lib/commands/make-icons.js
@@ -1,6 +1,6 @@
 const Command          = require('./-command');
 const iconTask         = require('splicon/src/icon-task');
-const getPlatforms     = require('../targets/cordova/utils/get-platforms');
+const getPlatforms     = require('../targets/cordova/utils/get-platforms-sync');
 const logger           = require('../utils/logger');
 
 const includes         = require('lodash').includes;
