+++ lib/commands/make-splashes.js
@@ -1,6 +1,6 @@
 const Command          = require('./-command');
 const splashTask       = require('splicon/src/splash-task');
-const getPlatforms     = require('../targets/cordova/utils/get-platforms');
+const getPlatforms     = require('../targets/cordova/utils/get-platforms-sync');
 const logger           = require('../utils/logger');
 
 const includes         = require('lodash').includes;
