+++ lib/targets/cordova/utils/get-platforms-sync.js
