+++ node-tests/unit/commands/make-icons-test.js
@@ -32,7 +32,7 @@ describe('Make Icons Command', function() {
     var logger;
 
     beforeEach(function() {
-      let getPlatforms = '../../../lib/targets/cordova/utils/get-platforms';
+      let getPlatforms = '../../../lib/targets/cordova/utils/get-platforms-sync';
       td.replace(getPlatforms, function() {
         return addedPlatforms;
       });
